Make turn count long again

+++ source/src/main/java/org/pintoschneider/void_of_the_unfathomable/game/scenes/InGame.java
@@ -381,7 +381,7 @@ public final class InGame implements Scene {
      *
      * @return The current turn count.
      */
-    public int turnCount() {
+    public long turnCount() {
         return turnManager.turnCount();
     }
 }
@@ -485,7 +485,7 @@ final class TurnManager {
      *
      * @return The turn count.
      */
-    public int turnCount() {
+    public long turnCount() {
         return turnCount;
     }
